requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10
pymupdf>=1.22.5
faiss-cpu>=1.7.4
streamlit>=1.26.0
python-dotenv>=1.0.0
openai>=1.1.1
tiktoken>=0.5.1